Restore stdcall-stripping for 32-bit import libs
This has gotten messy, as some .defs feature stdcall-mangled functions
as well as C++ ones, and some hide stdcall-mangled functions via `DATA`
suffix, so that we have to handle global variables with `@` characters
etc.
I switched from C to D, as `pragma(mangle)` is a big help when dealing
with arbitrary identifiers.