Change approach for weak symbols
Ignore the aliases in the MinGW-w64 .def files altogether; they are
not just incomplete, but sometimes also wrong (e.g.,
https://github.com/ldc-developers/ldc/issues/2903; there were more of
these after I checked against msvcrt.lib).

I decided to write a little tool to extract these aliases from the MS
libs. The resulting output for VS 2015 oldnames.lib and msvcrt.lib is
included (as well as the tool) as input files for the generated linker
directives in oldnames.lib.