Add a template to build releases on Github Actions.
This action is based off the dmd nightlies but also includes some
required changes to make it available as a reuseable workflow. This
allows actual workflows to call this template, e.g. for pull requests
against this repository.

Also includes a test action that uses the new action & verifies the
packaged compiler by building and running some basic program.

Note: The windows checks are partially disabled due to an upstream issue
      (invalid `sc.ini` due to the Win32 OMF => MSCOFF change)