Refactor release validation into a dedicated script
The checks were initially repeated for FreeBSD and the other platforms
because there wasn't much common code for both `step`s. But the upcoming
tests for the other bundled tools would require a lot of duplication.

This moves the code into a common script that is called by both steps