Fix Issue 22212 - dmd version has -dirty suffix on windows
Typically, git-clone probes the filesystem to see if it handles the executable bit correctly and the `core.fileMode` variable is automatically set as necessary. However when building dmd releases using vagrant box method, git-clone is done on a Linux host (i.e a filesystem that handles the filemode correctly), then scp'd to a Windows VM (i.e an environment that loses the filemode). So when building Windows binaries, always ensure this variable is set to `false`.